[feat][rate-limit] add header draft response
